app/core/admin.py
from django.contrib import admin
from django.contrib.auth.admin import UserAdmin

from .models import Member, UserProfile, User


class UserProfileInline(admin.StackedInline):
    model = UserProfile
    can_delete = False
    readonly_fields = ('uid', 'created_at', 'modified_at')
    verbose_name_plural = 'profile'
    fk_name = 'user'

class CustomUserAdmin(UserAdmin):
    inlines = (UserProfileInline, )

    def get_inline_instances(self, request, obj=None):
        if not obj:
            return list()
        return super(CustomUserAdmin, self).get_inline_instances(request, obj)

admin.site.register(Member, CustomUserAdmin)