requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.txt
#
--extra-index-url https://www.piwheels.org/simple

django-allauth
beautifulsoup4==4.6.0
celery==4.3.0
django-celery-beat==1.5.0
django-environ==0.4.5
django-extensions==2.2.5
django-redis==4.11.0
djangorestframework==3.11.0
pillow==6.2.1             # via wagtail
psycopg2
pytz==2019.3
redis==3.3.11
requests==2.22.0
gunicorn